+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- [#1925](https://github.com/teambit/bit/issues/1925) fix Angular non-relative paths from decorators
+
 ## [2.1.1-dev.1] - 2019-08-12
 
 - remove angular dependencies. use typescript compiler to parse Angular Decorators